Added collision
The game now has working collisions, implemented with the help of a
bounding box around the main character consisting of a Rectangle. The
"hard surfaces" were designed in Tiled and parsed into c++ with
tinyXML2. The collision-system works by comparing whether the object's
Rectangle overlaps with another one (the player's one for now).